IMethodValue to the separate file
