+++ src/IMethodValue.ts
@@ -1,0 +1,17 @@
+import * as vscode from "vscode";
+
+export interface IMethodValue {
+
+    name: string;
+
+    // начало
+    line: number;
+    // конец процедуры
+    endline: number;
+
+    filename: string;
+
+    description?: string;
+
+    kind?: vscode.CompletionItemKind;
+}
